working cart and database setup

+++ src/main/java/dat/backend/model/entities/Bottom.java
@@ -4,9 +4,9 @@ public class Bottom
 {
     private int id;
     private String name;
-    private int price;
+    private double price;
 
-    public Bottom(int id, String name, int price) {
+    public Bottom(int id, String name, double price) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -20,7 +20,7 @@ public class Bottom
         return name;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 }
