+++ src/main/java/dat/backend/model/entities/Top.java
@@ -4,9 +4,9 @@ public class Top
 {
     private int id;
     private String name;
-    private int price;
+    private double price;
 
-    public Top(int id, String name, int price) {
+    public Top(int id, String name, double price) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -21,7 +21,7 @@ public class Top
         return name;
     }
 
-    public int getPrice() {
+    public double getPrice() {
         return price;
     }
 }
